+++ src/server/renderer/dev.ts
@@ -38,7 +38,7 @@ export const enableDevRenderer = async (app: Express, cache: CacheClient) => {
           resolveTemplate({
             template,
             appHTML: redered.html,
-            metas: redered.metas,
+            heads: redered.heads,
             scripts: redered.scripts
           })
         )
@@ -49,7 +49,7 @@ export const enableDevRenderer = async (app: Express, cache: CacheClient) => {
         resolveTemplate({
           template,
           appHTML: redered.html,
-          metas: redered.metas,
+          heads: redered.heads,
           scripts: redered.scripts
         })
       )